Core google charts now use a core printer
