+++ views/chart.html
@@ -6,5 +6,8 @@
 	</div>
 </div>
 
-
+{{#unless alias}}
+	<h2>No queries set up yet.</h2>
+	<p>Visit <a href="https://docs.google.com/spreadsheets/d/1jH15yE5T6omD-B58UJfu1y4j8qN4QIs17u-52_Jkw7M/edit#gid=0">THE SPREADSHEET OF DESTINY</a> to set some up.
+{{/unless}}
